+++ GeniusTests/Extensions/Genius/MaintenanceCheck.swift
@@ -15,8 +15,8 @@ extension MaintenanceCheck where ValueWrapper == SyncValueWrapper<Value> {
 			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : nil
 		} else if let value = value.optional {
 			value >= requirement ? // swiftlint:disable:this void_function_in_ternary
-				Symbol(.checkmark, color: .green, label: .enabled) :
-				Symbol(.xmark, color: .red, label: .disabled)
+				Symbol(.checkmark, color: .green, label: .passed) :
+				Symbol(.xmark, color: .red, label: .failed)
 		} else {
 			Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ?
 				Symbol(.questionmark, color: .red, label: .unknown) : nil