+++ GeniusTests/Extensions/Genius/Symbol.swift
@@ -11,8 +11,8 @@ import SwiftUI
 extension Symbol: @retroactive Equatable {
 
 	static let notAvailable = Self(.minus, color: .primary, label: .notAvailable)
-	static let disabled = Self(.xmark, color: .red, label: .disabled)
-	static let enabled = Self(.checkmark, color: .green, label: .enabled)
+	static let failed = Self(.xmark, color: .red, label: .failed)
+	static let passed = Self(.checkmark, color: .green, label: .passed)
 	static let unknown = Self(.questionmark, color: .red, label: .unknown)
 
 	static func == (lhs: Self, rhs: Self) -> Bool {